refactor: split GrpcRegistration.java (920 lines) into 4 focused files
Split the monolithic GrpcRegistration.java into:
- GrpcRegistration.java (~47 lines): Server lifecycle and constructor only
- FreeplaneGrpcService.java (~650 lines): Extracted gRPC service implementation
- NodeUtils.java (~50 lines): Static node utility methods
- JsonHelper.java (~140 lines): JSON-to-mindmap helper methods

Preserves all functional behavior, logging, and gRPC API semantics.
No changes to proto files, Activator.java, or build configuration.
